Lookup API endpoints at session creation
Previously we would lookup the API endpoints in a lazy way when
they are first needed. I'm not sure why, since they would have to
be fetched to do anything useful with the session anyway.

Doing it directly on session creation simplifies things and leads
to more expected behaviour in the error case.