Refactor requirements
The unit test requirements are quite heavy (Varda), so we prefer
to separate them from the main requirements.

Also, as this is a PyPI published package, we don't want to rely
on specific versions of our dependencies, so we get rid of the
main `requirements.txt` file (`setup.py` is enough).