Field definitions on resources
Using a metaclass approach similar to what Django uses for its
model fields, we now define fields on resource classes with field
classes.

This allows for more flexibility in field types and conversions
of their values from the API to Python and vice versa. It also
makes for better Sphinx generated API docs and discoverability
of the Python API (tab completion and docstrings).